+++ app/src/main/java/com/ringov/yamblzweather/Config.java
@@ -13,5 +13,4 @@ public class Config {
         public static final String VERSION_PATH = "data/2.5/";
         public static final String KEY = "b3f1079128fa7ee84d98e92fecc133fa";
     }
-
 }
