+++ app/src/main/java/com/ringov/yamblzweather/di/app/AppComponent.java
@@ -4,9 +4,9 @@ import com.ringov.yamblzweather.di.about.AboutModule;
 import com.ringov.yamblzweather.di.settings.SettingsModule;
 import com.ringov.yamblzweather.di.weather.WeatherModule;
 import com.ringov.yamblzweather.model.background_service.WeatherUpdateJob;
-import com.ringov.yamblzweather.ui.settings.SettingsFragment;
-import com.ringov.yamblzweather.viewmodel.about.AboutViewModel;
-import com.ringov.yamblzweather.viewmodel.weather.WeatherViewModel;
+import com.ringov.yamblzweather.presentation.ui.settings.SettingsFragment;
+import com.ringov.yamblzweather.presentation.ui.about.AboutViewModel;
+import com.ringov.yamblzweather.presentation.ui.weather.WeatherViewModel;
 
 import javax.inject.Singleton;
