+++ app/src/main/java/com/ringov/yamblzweather/model/Converter.java
@@ -3,8 +3,8 @@ package com.ringov.yamblzweather.model;
 import com.ringov.yamblzweather.model.db.data.DBWeather;
 import com.ringov.yamblzweather.model.internet.data.Weather;
 import com.ringov.yamblzweather.model.internet.data.ResponseWeather;
-import com.ringov.yamblzweather.viewmodel.data.WeatherCondition;
-import com.ringov.yamblzweather.viewmodel.data.UIWeather;
+import com.ringov.yamblzweather.presentation.data.WeatherCondition;
+import com.ringov.yamblzweather.presentation.data.UIWeather;
 
 import java.util.List;
