+++ app/src/main/java/com/ringov/yamblzweather/presentation/base/StateData.java
@@ -1,4 +1,4 @@
-package com.ringov.yamblzweather.viewmodel.base;
+package com.ringov.yamblzweather.presentation.base;
 
 /**
  * Created by ringov on 21.07.17.
