+++ app/src/main/java/com/ringov/yamblzweather/presentation/ui/MainViewUpdater.java
@@ -1,11 +1,12 @@
-package com.ringov.yamblzweather;
+package com.ringov.yamblzweather.presentation.ui;
 
-import com.ringov.yamblzweather.viewmodel.data.UIWeather;
+import com.ringov.yamblzweather.presentation.data.UIWeather;
 
 /**
  * Created by ringov on 16.07.17.
  */
 
 public interface MainViewUpdater {
+
     void onWeatherUpdate(UIWeather weather);
 }
