+++ app/src/main/java/com/ringov/yamblzweather/presentation/ui/settings/SettingsFragment.java
@@ -1,4 +1,4 @@
-package com.ringov.yamblzweather.ui.settings;
+package com.ringov.yamblzweather.presentation.ui.settings;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
