+++ app/src/main/java/com/ringov/yamblzweather/presentation/ui/weather/WeatherStateData.java
@@ -1,0 +1,10 @@
+package com.ringov.yamblzweather.presentation.ui.weather;
+
+import com.ringov.yamblzweather.presentation.base.StateData;
+
+/**
+ * Created by ringov on 21.07.17.
+ */
+
+public class WeatherStateData extends StateData {
+}
