+++ app/src/main/java/com/ringov/yamblzweather/presentation/ui/weather/WeatherView.java
@@ -1,0 +1,10 @@
+package com.ringov.yamblzweather.presentation.ui.weather;
+
+import com.ringov.yamblzweather.presentation.base.BaseView;
+
+/**
+ * Created by ringov on 21.07.17.
+ */
+
+public interface WeatherView extends BaseView {
+}
