+++ app/src/main/java/com/ringov/yamblzweather/routing/ScreenRouter.java
@@ -3,13 +3,6 @@ package com.ringov.yamblzweather.routing;
 import android.support.annotation.IdRes;
 import android.support.v4.app.FragmentManager;
 
-import com.ringov.yamblzweather.ui.about.AboutFragment;
-import com.ringov.yamblzweather.R;
-import com.ringov.yamblzweather.ui.weather.WeatherFragment;
-import com.ringov.yamblzweather.ui.settings.SettingsFragment;
-
-import java.util.HashMap;
-
 /**
  * Created by ringov on 07.07.17.
  */
