--- app/src/main/java/com/ringov/yamblzweather/ui/weather/WeatherStateData.java
@@ -1,10 +1,0 @@
-package com.ringov.yamblzweather.ui.weather;
-
-import com.ringov.yamblzweather.viewmodel.base.StateData;
-
-/**
- * Created by ringov on 21.07.17.
- */
-
-public class WeatherStateData extends StateData {
-}
