--- app/src/main/java/com/ringov/yamblzweather/ui/weather/WeatherView.java
@@ -1,10 +1,0 @@
-package com.ringov.yamblzweather.ui.weather;
-
-import com.ringov.yamblzweather.ui.base.BaseView;
-
-/**
- * Created by ringov on 21.07.17.
- */
-
-public interface WeatherView extends BaseView {
-}
